Remove ExcludeUrlFilter and simply use ExcludePathFilter
