+++ bootstrap-core/src/main/java/com/navercorp/pinpoint/bootstrap/config/ExcludePathFilter.java
@@ -26,6 +26,7 @@ import java.util.Collections;
 import java.util.List;
 
 /**
+ * @author emeroad
  * @author HyunGil Jeong
  */
 public class ExcludePathFilter implements Filter<String> {
