--- bootstrap-core/src/main/java/com/navercorp/pinpoint/bootstrap/config/ExcludeUrlFilter.java
@@ -1,38 +1,0 @@
-/*
- * Copyright 2014 NAVER Corp.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.navercorp.pinpoint.bootstrap.config;
-
-/**
- * @author emeroad
- */
-public class ExcludeUrlFilter extends ExcludePathFilter {
-
-
-
-    public ExcludeUrlFilter(String excludePathFormatString) {
-        super(excludePathFormatString);
-    }
-
-    @Override
-    public String toString() {
-        final StringBuilder sb = new StringBuilder("ExcludeUrlFilter{");
-        sb.append("excludeUrlMatchers=").append(excludePathMatchers);
-        sb.append('}');
-        return sb.toString();
-    }
-}
-
